+++ R/aims_molybdenum_marine.R
@@ -19,13 +19,14 @@
 #'\item{Species}{The species names name (chr).}
 #'\item{Test_endpoint}{Endpoint statistic, EC10, NEC etc (chr).}
 #'\item{Timeframe}{Exposure timeframe basis of the value: "chronic" or "short_term". (chr).}
-#'\item{Toxicity_measure}{Type of toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{Type of toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #' }
 #' 
 #' @name aims_molybdenum_marine
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 14 rows and 10 columns.
+#' \code{data.frame}) with 14 rows and 11 columns.
 #' @keywords datasets
 #' @examples
 #'