+++ R/anzg_data.R
@@ -59,13 +59,14 @@
 #'\item{Genus}{The Genus name (chr).}
 #'\item{Group}{The taxonomic group (chr).}
 #'\item{Life_stage}{Life stage of the test organism (chr).}
-#'\item{Medium}{The medium - freshwater or marine water (chr).}
+#'\item{Medium}{The test medium: "Freshwater", "Marine", or one of the hardness variants "Soft freshwater", "Moderate freshwater", "Hard freshwater" (chr).}
 #'\item{Notes}{Other notes (chr).}
 #'\item{Phylum}{The Phylum name (chr).}
 #'\item{Species}{The species binomial name (chr).}
 #'\item{Test_endpoint}{The test endpoint measure (chr).}
 #'\item{Timeframe}{Exposure timeframe basis of the value: "chronic" or "short_term". (chr).}
-#'\item{Toxicity_measure}{The toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{The toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #'
 #' 
 #' Where toxicity measure is not a chronic NEC, EC10 or NOEC value,
@@ -77,7 +78,7 @@
 #' @name anzg_data
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 592 rows and 13 columns.
+#' \code{data.frame}) with 592 rows and 14 columns.
 #' @keywords datasets
 #' @examples
 #'