+++ R/anzg_simazine_marine.R
@@ -25,7 +25,8 @@
 #'\item{Species}{The species binomial name (chr).}
 #'\item{Test_endpoint}{The test endpoint measure (chr).}
 #'\item{Timeframe}{Exposure timeframe basis of the value: "chronic" or "short_term". (chr).}
-#'\item{Toxicity_measure}{The toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{The toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #'
 #' 
 #' Where toxicity measure is not a chronic NEC, EC10 or NOEC value,
@@ -37,7 +38,7 @@
 #' @name anzg_simazine_marine
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 14 rows and 10 columns.
+#' \code{data.frame}) with 14 rows and 11 columns.
 #' @keywords datasets
 #' @examples
 #'