+++ R/csiro_data.R
@@ -18,17 +18,18 @@
 #' 
 #' \describe{ 
 #'\item{Chemical}{The chemical name (chr).}
-#'\item{Conc}{The chemical concentration (dbl).}
+#'\item{Conc}{The chemical concentration in micrograms per Litre (dbl).}
 #'\item{Domain}{Tropical, temperate or other filter (chr).}
 #'\item{Duration}{Test duration (chr).}
 #'\item{Group}{Taxonomic grouping information (chr).}
 #'\item{Life_stage}{Life stage of the test organism (chr).}
-#'\item{Medium}{The medium - fresh or marine water (chr).}
+#'\item{Medium}{The test medium: "Freshwater" or "Marine" (chr).}
 #'\item{Notes}{Other notes (chr).}
 #'\item{Species}{The species names name (chr).}
 #'\item{Test_endpoint}{Endpoint statistic, EC10, NEC etc (chr).}
 #'\item{Timeframe}{Exposure timeframe basis of the value: "chronic" or "short_term". (chr).}
-#'\item{Toxicity_measure}{Type of toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{Type of toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #'
 #' 
 #' Where toxicity measure is not a chronic NEC, EC10 or NOEC value,
@@ -42,7 +43,7 @@
 #' @name csiro_data
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 91 rows and 12 columns.
+#' \code{data.frame}) with 91 rows and 13 columns.
 #' @keywords datasets
 #' @examples
 #'